+++ dom/src/main/java/org/linguafranca/pwdb/kdbx/dom/DomDatabaseWrapper.java
@@ -177,4 +177,6 @@ public class DomDatabaseWrapper extends AbstractDatabase<DomDatabaseWrapper, Dom
         DomHelper.touchElement("DatabaseDescriptionChanged", dbMeta);
         setDirty(true);
     }
+
+
 }