+++ kdb/src/test/java/org/linguafranca/pwdb/kdb/KdbDatabaseTest.java
@@ -16,10 +16,13 @@
 
 package org.linguafranca.pwdb.kdb;
 
+import org.junit.Test;
 import org.linguafranca.pwdb.checks.BasicDatabaseChecks;
 import org.linguafranca.pwdb.Database;
 
 import java.io.IOException;
+
+import static org.junit.Assert.assertFalse;
 
 /**
  * @author Jo
@@ -32,5 +35,12 @@ public class KdbDatabaseTest  extends BasicDatabaseChecks {
 
     public KdbDatabaseTest() throws IOException {
         super();
+    }
+
+    @Test
+    public void supportedFunctionalityTest(){
+        assertFalse(database.supportsBinaryProperties());
+        assertFalse(database.supportsNonStandardPropertyNames());
+        assertFalse(database.supportsRecycleBin());
     }
 }