+++ simple/src/test/java/org/linguafranca/pwdb/kdbx/simple/SimpleEntryTest.java
@@ -16,7 +16,7 @@
 
 package org.linguafranca.pwdb.kdbx.simple;
 
-import org.linguafranca.pwdb.checks.EntryChecks;
+import org.linguafranca.pwdb.checks.BinaryPropertyChecks;
 import org.linguafranca.pwdb.Database;
 import org.linguafranca.pwdb.kdbx.KdbxCreds;
 import org.linguafranca.pwdb.Credentials;
@@ -28,7 +28,7 @@ import java.io.OutputStream;
 /**
  * @author jo
  */
-public class SimpleEntryTest extends EntryChecks {
+public class SimpleEntryTest extends BinaryPropertyChecks {
 
     public SimpleEntryTest() {
         InputStream inputStream = getClass().getClassLoader().getResourceAsStream("Attachment.kdbx");